+++ src/main/java/com/qingfengzhuyue/forum/model/QuestionExample.java
@@ -674,52 +674,52 @@ public class QuestionExample {
             return (Criteria) this;
         }
 
-        public Criteria andShieldEqualTo(Long value) {
+        public Criteria andShieldEqualTo(Integer value) {
             addCriterion("shield =", value, "shield");
             return (Criteria) this;
         }
 
-        public Criteria andShieldNotEqualTo(Long value) {
+        public Criteria andShieldNotEqualTo(Integer value) {
             addCriterion("shield <>", value, "shield");
             return (Criteria) this;
         }
 
-        public Criteria andShieldGreaterThan(Long value) {
+        public Criteria andShieldGreaterThan(Integer value) {
             addCriterion("shield >", value, "shield");
             return (Criteria) this;
         }
 
-        public Criteria andShieldGreaterThanOrEqualTo(Long value) {
+        public Criteria andShieldGreaterThanOrEqualTo(Integer value) {
             addCriterion("shield >=", value, "shield");
             return (Criteria) this;
         }
 
-        public Criteria andShieldLessThan(Long value) {
+        public Criteria andShieldLessThan(Integer value) {
             addCriterion("shield <", value, "shield");
             return (Criteria) this;
         }
 
-        public Criteria andShieldLessThanOrEqualTo(Long value) {
+        public Criteria andShieldLessThanOrEqualTo(Integer value) {
             addCriterion("shield <=", value, "shield");
             return (Criteria) this;
         }
 
-        public Criteria andShieldIn(List<Long> values) {
+        public Criteria andShieldIn(List<Integer> values) {
             addCriterion("shield in", values, "shield");
             return (Criteria) this;
         }
 
-        public Criteria andShieldNotIn(List<Long> values) {
+        public Criteria andShieldNotIn(List<Integer> values) {
             addCriterion("shield not in", values, "shield");
             return (Criteria) this;
         }
 
-        public Criteria andShieldBetween(Long value1, Long value2) {
+        public Criteria andShieldBetween(Integer value1, Integer value2) {
             addCriterion("shield between", value1, value2, "shield");
             return (Criteria) this;
         }
 
-        public Criteria andShieldNotBetween(Long value1, Long value2) {
+        public Criteria andShieldNotBetween(Integer value1, Integer value2) {
             addCriterion("shield not between", value1, value2, "shield");
             return (Criteria) this;
         }
